+++ include/aarchvm/soc.hpp
@@ -12,6 +12,7 @@
 #include "aarchvm/pl050_kmi.hpp"
 #include "aarchvm/perf_types.hpp"
 #include "aarchvm/ram.hpp"
+#include "aarchvm/rtc_pl031.hpp"
 #include "aarchvm/uart_pl011.hpp"
 
 #include <cstddef>
@@ -119,6 +120,8 @@ private:
   static constexpr std::uint64_t kKmiSize = 0x1000;
   static constexpr std::uint64_t kPerfBase = 0x09020000;
   static constexpr std::uint64_t kPerfSize = 0x1000;
+  static constexpr std::uint64_t kRtcBase = 0x09030000;
+  static constexpr std::uint64_t kRtcSize = 0x1000;
   static constexpr std::uint64_t kBlockBase = 0x09040000;
   static constexpr std::uint64_t kBlockSize = 0x1000;
   static constexpr std::uint64_t kGicBase = 0x08000000;
@@ -199,6 +202,7 @@ private:
   std::shared_ptr<UartPl011> uart_;
   std::shared_ptr<Pl050Kmi> kmi_;
   std::shared_ptr<PerfMailbox> perf_mailbox_;
+  std::shared_ptr<RtcPl031> rtc_;
   std::shared_ptr<BlockMmio> block_mmio_;
   std::shared_ptr<GicV3> gic_;
   std::shared_ptr<GenericTimer> timer_;